refactor: extract *MenuScreen components from 9 root game files (#293)
Each root game component now delegates its menu phase to a dedicated
*MenuScreen component, making root files pure phase-routers.
